% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellate_functions.R
\name{get_cnstl}
\alias{get_cnstl}
\title{Get constellate papers based on search criteria}
\usage{
get_cnstl(
  cnstl_dir,
  fields,
  target_journals = NULL,
  text_search = NULL,
  use_stable_names = FALSE,
  min_year = NULL,
  max_year = NULL,
  limit = NULL,
  return_query = FALSE
)
}
\arguments{
\item{cnstl_dir}{File path for directory with constellate data (Arrow will
grab all files within this folder and subfolders).}

\item{fields}{Vector of fields to return. Using * to return all fields
may result in errors.}

\item{text_search}{Vector of strings to match in texts. Texts will only
be returned that contain at least one of the strings in this vector. If NULL,
all papers will be returned. Search is not case sensitive.}

\item{use_stable_names}{If TRUE, a stable_name field will be returned that
lists the most recent name of journals that have changed names over time. If
target_journals has been specified, all journals that share the target
journals' stable names will be added to your search.}

\item{min_year}{If provided, results will exclude papers from before min_year}

\item{max_year}{If provided, results will exclude papers from after max_year}

\item{limit}{If provided, will return only this number of papers (for testing)}

\item{return_query}{If TRUE, return the query instead of results. To be used
when you want to add to the base query}

\item{target_jounals}{Vector of target journal names. If NULL, papers
from all journals will be returned.}
}
\value{
A dataframe with results of query
}
\description{
Get constellate papers based on search criteria
}
