% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellate_functions.R
\name{journal_count}
\alias{journal_count}
\title{Get journal-level (or journal-by-year) counts of how many papers
contain matching text}
\usage{
journal_count(cnstl_dir, text_search, by_year = FALSE)
}
\arguments{
\item{cnstl_dir}{File path for directory with constellate data (Arrow will
grab all files within this folder and subfolders)}

\item{text_search}{Vector of strings to match in texts. Texts will only
be returned that contain at least one of the strings in this vector.}

\item{by_year}{If TRUE, results will be aggregated at the
journal-by-year level}
}
\value{
A dataframe with counts of matching papers in each journal
(or journal-by-year if by_year is set to TRUE)
}
\description{
Get journal-level (or journal-by-year) counts of how many papers
contain matching text
}
