% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellate_functions.R
\name{list_cnstl_fields}
\alias{list_cnstl_fields}
\title{List all fields in Constellate Data}
\usage{
list_cnstl_fields(cnstl_dir)
}
\arguments{
\item{cnstl_dir}{File path for directory with constellate data (Arrow will
grab all files within this folder and subfolders)}
}
\value{
Character vector containing the field names in the Constellate data
}
\description{
List all fields in Constellate Data
}
