% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellate_functions.R
\name{plot_time_series_by_term}
\alias{plot_time_series_by_term}
\title{Plot term series of how frequently term is used in corpus}
\usage{
plot_time_series_by_term(cnstl_dir, text_search)
}
\arguments{
\item{cnstl_dir}{File path for directory with constellate data (Arrow will
grab all files within this folder and subfolders)}

\item{text_search}{Vector of strings to match in texts.}
}
\value{
A ggplot object (prints the plot as a side effect).
}
\description{
Plot term series of how frequently term is used in corpus
}
