% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellate_functions.R
\name{text_match_time_series}
\alias{text_match_time_series}
\title{Get annual level counts of how many papers match a term or set of terms}
\usage{
text_match_time_series(cnstl_dir, text_search)
}
\arguments{
\item{cnstl_dir}{File path for directory with constellate data (Arrow will
grab all files within this folder and subfolders)}

\item{text_search}{Vector of strings to match in texts.}
}
\value{
A dataframe with annual counts of how many times the specified terms
appear
}
\description{
Get annual level counts of how many papers match a term or set of terms
}
